requirements.txt
pystray
openai
whisper-mic
keyboard
requests
Pillow
pydub
simpleaudio
tkinter 
pyttsx3
speechrecognition
